+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2019-07-05
+
 ### Added
 - Segmented playback for videos, auto starting and pausing at the time specified
   ```typescript